update discard logic, change dates format to long

+++ config/initializers/active_admin.rb
@@ -153,7 +153,7 @@ ActiveAdmin.setup do |config|
   # To understand how to localize your app with I18n, read more at
   # https://github.com/svenfuchs/i18n/blob/master/lib%2Fi18n%2Fbackend%2Fbase.rb#L52
   #
-  config.localize_format = :short
+  config.localize_format = :long
 
   # == Setting a Favicon
   #
